+++ CHANGES.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## v. 0.6.0
+ * rewritten full iteration for table documents, using the column name as 
+   context
+ * fixes in YAML dump
+ * additional functions for loading documents from JSON/YAML
+
+
+## v. 0.5.1
+ * `DocumentChunk` method renamed as_dict() => asdict() for consistency
+   (with an alias to the old name for backwards compatibility)
+ * improved error handling of invalid config files
+ * fix: rawdoc script
+
+
 ## v. 0.5.0
  * added "extra" field to PiiEntity
  * added add_process_stage() method to PiiEntity